Fix a few overlooked spots where the coded uses 16-bit NSIDs. Chuck Tuffli had submitted a more thorough patch that I was unaware of when I did my work and this brings in the bits I missed from that patch.
PR: 220267
Submitted by: Chuck Tuffli


git-svn-id: svn+ssh://svn.freebsd.org/base/head@322998 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/cam/cam_ccb.h
@@ -631,7 +631,7 @@ struct ccb_pathinq_settings_sas {
 };
 
 struct ccb_pathinq_settings_nvme {
-	uint16_t nsid;		/* Namespace ID for this path */
+	uint32_t nsid;		/* Namespace ID for this path */
 };
 
 #define	PATHINQ_SETTINGS_SIZE	128
